refactor image entity to new format
