--- SportMap.DAL.Abstractions/IImageRepository.cs
@@ -1,8 +1,0 @@
-using DomainLayer.Entities;
-
-namespace SportMap.DAL.Abstractions
-{
-    public interface IImageRepository : IRepository<ImageData>
-    {
-    }
-}
