+++ SportMap.DAL.Abstractions/Repositories/IImageRepository.cs
@@ -1,0 +1,6 @@
+using DomainLayer.Entities;
+
+namespace SportMap.DAL.Abstractions.Repositories
+{
+    public interface IImageRepository : IRepository<ImageData>;
+}
