+++ SportMap.DAL/Repositories/ImageRepository.cs
@@ -1,13 +1,11 @@
 using DomainLayer.Entities;
 using Microsoft.Extensions.Logging;
-using SportMap.DAL.Abstractions;
+using SportMap.DAL.Abstractions.Repositories;
 using SportMap.DAL.Common;
 using SportMap.DAL.DataContext;
 
-namespace SportMap.DAL.DataAccess
+namespace SportMap.DAL.Repositories
 {
     public class ImageRepository(AppDbContext context, ILogger<ImageRepository> logger)
-        : BaseRepository<ImageData>(context, logger, context.Images), IImageRepository
-    {
-    }
+        : BaseRepository<ImageData>(context, logger, context.Images), IImageRepository;
 }
